Make super-save-auto-save-when-idle buffer local
In some cases idle save should be shutdown. For example,
if a function such as 'gofmt-before-save or
'go-remove-unused-imports is added in before-save-hook, then
every idle-save will format the code. Some lines/characters
might be deleted if the format tool believes it were unused,
which may not be true. Make this variable buffer local will
solve this problem.